Debugging drag & drop and highlight; added snap guidelines
Prev issues: when dragging and dropping blocks, it had left behind the highlighted box. Additionally, when dragging and dropping over other blocks, it would stack blocks horizontally instead of vertically.

The reason for the the issues was due to the blocks not being able to handle child elements properly when dragging, dropping, and deleting.

Fixes: allowed for dragging, dropping, and deleting to allow for child elements to be brought along together

Added snap guidelines: added a blue line when dragging blocks to indicate where the block's new location will be